pgn-etl/pyproject.toml
[tool.poetry]
name = "pgn-etl"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
"chess.com" = "^1.7.6"
prefect = "2.0a12"
greenlet = "^1.1.2"
pydantic = "^1.9.0"
boto3 = "^1.21.5"
pandas = "^1.4.1"
python-chess = "^1.999"
ipython = "^8.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
